PvP Conway's Pre-Beta v1.0
Prefab buttons are now fully functional:
    When clicked, a pattern is attached to your cursor that can be placed anywhere it fits with left click, or cleared with right click
    The pattern can be rotated with 'r' and flipped vertically with 'f' and horizontally with 't'

While time is off, if you make a mistake you can now undo one action with ctrl + z

This meets all of the gameplay functionality I was planning for the Beta version of the game. All that is left is creating levels, prettying things up, and making a basic tutorial/guide. As such, the version has been renamed to Pre-Beta.